refactor, build: 카카오 인증키처리방식 변경
