+++ App/Sources/Presentation/Map/Cells/SearchAddressCollectionViewCell.swift
@@ -7,6 +7,8 @@
 //
 
 import UIKit
+
+import CODomain
 import SnapKit
 
 class SearchAddressCollectionViewCell: UICollectionViewCell {
