+++ Auth/COAuth/Sources/Auth.swift
@@ -1,0 +1,14 @@
+//
+//  ThirdParty.swift
+//  COAuth
+//
+//  Created by sean on 2022/09/28.
+//
+
+import Foundation
+
+public struct Auth {
+  public struct ThirdParty {
+    public static let kakao: String = "KakaoAK b1cc03bcc27ce1a6b16c76e5132712ac"
+  }
+}
