+++ Features/Sign/Project.swift
@@ -7,7 +7,7 @@ let isExcludedFramework = (excludedFramework == "TRUE")
 
 let sign = Project.feature(
   name: "Sign",
-  products: [.framework(.dynamic), .demoApp, .unitTests],
+  products: [.framework(.static), .demoApp, .unitTests],
   isExcludedFramework: isExcludedFramework,
   infoExtension: [
     "LSApplicationQueriesSchemes": .array(
@@ -15,11 +15,12 @@ let sign = Project.feature(
     )
   ],
   dependencies: [
+    .project(target: "COAuth", path: .relativeToRoot("Auth/COAuth")),
     .project(target: "COCommonUI", path: .relativeToRoot("UI/COCommonUI")),
     .project(target: "COFoundation", path: .relativeToRoot("Core/COFoundation")),
     .project(target: "CONetwork", path: .relativeToRoot("Core/CONetwork")),
-    .project(target: "CODomain", path: .relativeToRoot("Domain/CODomain")),
     .project(target: "COThirdParty", path: .relativeToRoot("Core/COThirdParty")),
+    .project(target: "CODomain", path: .relativeToRoot("Domain/CODomain"))
   ],
   externalDependencies: [
     .xcframework(path: .CocoaPods.Framework.naverLogin)
